refactor: 리팩토링 사용자 입력 유효성 검사에 필요한 함수 유틸리티 함수로 분류
