+++ modules/inputValidation.js
@@ -1,26 +1,12 @@
-import { MESSAGE, RANGE_NUMBER } from './constant';
-
-const checkCarNamesHaveBlank = (carNames) =>
-  carNames.includes('') || carNames.includes(' ');
-
-const checkCarNameIsOverFive = (carNames) => {
-  const inputHaveOverFive = carNames.find(
-    (carName) => carName.length > RANGE_NUMBER.carName
-  );
-  return inputHaveOverFive;
-};
-
-const checkCarNameWithBlank = (carNames) => {
-  const inputWithBlank = carNames.filter((car) => car.includes(' '));
-  return inputWithBlank.length !== 0;
-};
-
-const checkCarNamesAreDuplicated = (carNames) => {
-  const carNumber = carNames.length;
-  const nonDuplicatedCarNames = [...new Set(carNames)];
-
-  return carNumber !== nonDuplicatedCarNames.length;
-};
+import { MESSAGE } from './constant';
+import {
+  checkCarNamesHaveBlank,
+  checkCarNameIsOverFive,
+  checkCarNameWithBlank,
+  checkCarNamesAreDuplicated,
+  checkTryNumberWithBlank,
+  checkTryNumberIsCharacter,
+} from '../modules/utils';
 
 const checkCarNamesAreValid = (userInput) => {
   const input = userInput.split(',');
@@ -40,14 +26,6 @@ const checkCarNamesAreValid = (userInput) => {
   }
   return input;
 };
-
-const checkTryNumberWithBlank = (userInput) => {
-  const inputHaveBlank = userInput.includes(' ');
-  return inputHaveBlank;
-};
-
-const checkTryNumberIsCharacter = (userInput) =>
-  Number.isNaN(Number(userInput)) || Number(userInput) === 0;
 
 const checkTryNumberIsValid = (userInput) => {
   const inputHaveBlank = checkTryNumberWithBlank(userInput);
